requirements.txt
pandas
matplotlib
seaborn
jupyterlab
python-dotenv
requests
